Fix explanation for UsersTableSeeder
The explanation given to the `UsersTableSeeder` seemed to be interchanged in terms of order of code execution.
In reality, the *50 random users* will be created first along with 5 articles each, before creating an *additional user* with an email of `me@mygraphqlapp.com` which is an opposite of what the original explanation gave. So in this PR I just interchanged the order of the code (which is much easier) to match the explanation.